In progress the development of I2C communication with dsNav board. The base concepts got from barometer branch from dberroya.
Implemented and tested the switch of the clock to 40Mips instead of 16. Concepts got from PilotQuad branch from Mark Whitehom.
Implemented and tested the testPoint function to test correct timing of critical routines.
Implemented and tested the internal conversion from TOW + WEEKS number to date and time. (to be used carefully).
Added command T to send date and time params to the console